+++ RightTriangle.cpp
@@ -3,7 +3,7 @@
 #include "RightTriangle.h"
 #include "Canvas.h"
 
-RightTriangle::RightTriangle(int b, char m_pen, std::string m_name) : Triangle(b, b, m_name, m_pen) {}
+RightTriangle::RightTriangle(size_t b, char m_pen, std::string m_name) : Triangle(b, b, m_name, m_pen) {}
 
 double RightTriangle::perimeterGeo() const { return (2 + std::sqrt(2))*height; }
