+++ Shape.cpp
@@ -3,7 +3,7 @@
 
 size_t Shape::counter = 1; 
 
-Shape::Shape(int m_height, int m_width, std::string m_name, char m_pen) 
+Shape::Shape(size_t m_height, size_t m_width, std::string m_name, char m_pen) 
 {
     name = m_name;
     pen = m_pen;
@@ -19,17 +19,17 @@ char Shape::getPen() const { return pen; }
 
 int Shape::getId() const { return id; }
 
-int Shape::getHeight() const { return height; }
+size_t Shape::getHeight() const { return height; }
 
-int Shape::getWidth() const { return width;}
+size_t Shape::getWidth() const { return width;}
 
 void Shape::setName(const std::string& m_name) { name = m_name; }
 
 void Shape::setPen(const char& m_pen) { pen = m_pen; }
 
-void Shape::setHeight(const int& m_height) { height = m_height; }
+void Shape::setHeight(const size_t& m_height) { height = m_height; }
 
-void Shape::setWidth(const int& m_width) { width = m_width; }
+void Shape::setWidth(const size_t& m_width) { width = m_width; }
 
 std::string Shape::toString() const
 {
